feat:新增每日排程 (搭配檢查今天的爬蟲檔案並且強制讀檔更新的參數)
每日排程目前設定開啟時候會強制跑一次指令

+++ restaurant_app/models.py
@@ -23,14 +23,14 @@ class Restaurant(models.Model):
     name = models.CharField(max_length=255, verbose_name='餐廳名稱')
     hash_value = models.CharField(max_length=64, unique=True, blank=False, null=False, verbose_name='Hash 值')
     rating = models.FloatField(verbose_name='評分')
-    review_count = models.IntegerField(verbose_name='評論數')
-    address = models.CharField(max_length=255, verbose_name='地址')
+    review_count = models.IntegerField(verbose_name='評論數', default=0)
+    address = models.CharField(max_length=255, verbose_name='地址', blank=False, null=False)
     phone_number = models.CharField(max_length=20, null=True, blank=True, verbose_name='電話號碼')  # 新增電話號碼欄位
     average_spending = models.CharField(max_length=255, blank=True, null=True, verbose_name='平均消費')
     opening_hours = models.JSONField(blank=True, null=True, verbose_name='營業時間')
     services = models.JSONField(blank=True, null=True, verbose_name='提供服務')
-    latitude = models.FloatField(verbose_name='緯度')
-    longitude = models.FloatField(verbose_name='經度')
+    latitude = models.FloatField(verbose_name='緯度', blank=False, null=False)
+    longitude = models.FloatField(verbose_name='經度', blank=False, null=False)
     image_url = models.CharField(max_length=500, blank=True, null=True, verbose_name='圖片網址')
     google_url = models.URLField(max_length=1000, null=True, blank=True, verbose_name='google網址')  # Google Maps 連結欄位
     # 其他欄位
@@ -62,6 +62,9 @@ class Restaurant(models.Model):
         
     
     def updateBusinesshours(self):
+        if not self.opening_hours:
+            return
+        
         # 刪除當前餐廳的所有 Businesshours 以避免重複
         Businesshours.objects.filter(restaurant=self).delete()
